requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
jinja2==3.1.3
Pillow==10.2.0
moviepy==1.0.3
mutagen==1.47.0
sentence-transformers==2.2.2
torch>=1.9.0
numpy>=1.21.0
chromadb>=0.4.0 